Fixing small mistakes & cleaning edge_serving code
